Move the import logic to its own .js
